Daly Light Savings
This fix addresses an issue with how reports IDs are captured. Since the report starter endpoint doesn't return an ID of the new report, it's retrieved by looking at the report end date. This is a date-time data type with a time zone. This fix address subtracting an hour from the lookup when the end date is within the DST time zone.