Rename/Updating code
Renamed the file and changed inputs. Now take the path to where all the MAID files are stored.

returns a dictionary with the total count of blobs and for each blob it has the url (with sas) and the total number of maids within the blob.

Having issues with the blob.query.