Sales audience automation (#79)
* Async neighbors logic

Added asynchronous logic for the neighbors pull. Needs more work and for an Isaac touch

* async and multithreaded

* identifying eav audience sources

* primary data query

* apply data filter

* docstring and defaults

* init of neighbors logic

* new sales bind

* moving files to utils

* Major neighbors push

* variable name

* variable naming

* blob name

* blob naming

* missing data handling

* modify output logic

* typo

* Bulk update

The major set of work that's gone into the neighbors automation

* processing step outputType fix

got the correct sqlalchemy for getting the outputType title

* successful test

First push of a successful test to get us deviceids in the folder for step 1

* multiple neighbors output files

changed it so we're not combining all of the files together for each of the neighbors pulls, but instead doing one blob url for each

* date limitation sales addition

* owned location neighbors radius query

Major commit for the logic that takes the sales data, keeping only sales that are within X distance of any owned location of the client

* clarity

* Primary shift to new steps system

* Bulk update

For our functioning test on the new steps

* Logging

* Bugfix

fixed the outputlocation that lead to a bug causing the final onspot calls to error

* Graceful check

Little bit of a grace check beyond the get for if there's no processing

* Cleanup

Removed a lot of logging from the dev phase and renamed for clarity

* Data Filter custom handling

added a function that checks for a custom field and another that takes the custom field and turns it into the proper filter

* sales data query update

updated to handle more datafilters such as the custom fields that are the raw attribute names. also added a function that removes the days_back clause since that isn't actually a clause we need to use explicitly

* updated sales query

updated the sales data query so that it will get all necessary fields, cast properly, and make the datafilter where clause work properly. additionally reduced address info pull for speed and clarity since they're not all necessary

* update to handle custom filters

---------

Co-authored-by: Isaac Jessup <ibjessup@gmail.com>